+++ noodles-vcf/src/variant.rs
@@ -1,0 +1,3 @@
+//! Variant format.
+
+pub mod io;